feat(log): add compact format to log print
The record value reader used by the log print was a cache. This was not
usable for the compact record logger, since it requires all records as
an in memory collection (we need new objects for each record). Therefore
the metadata and loggedEvent objects are also created newly when reading
each individual record. For large logs, this may lead to large memory
usage.