feat: add key to incident details
The key of an incident is important information, especially for the list
command. Consider having some incident in the state which you're looking
up with zdb:

 zdb incident -p ./data/raft-partition/partitions/1/1runtime list | jq

Which then shows you all kinds of details about incidents, except the
key. However, the key is necessary to resolve the issue. So often you're
actually looking for that data.

Adding it is realatively easy, because in both the list and entity
command we already have the incident key available when we create the
details object.